requirements.txt
numpy>=1.24.0
pandas>=2.0.0
ta>=0.10.0  # Technical analysis library
ccxt>=4.0.0  # Cryptocurrency exchange library
python-binance>=1.0.17  # Binance API
yfinance>=0.2.0  # Yahoo Finance API for S&P 500 data
fastapi>=0.100.0  # For API backend
uvicorn>=0.22.0  # ASGI server
python-dotenv>=1.0.0
scikit-learn>=1.3.0
plotly>=5.15.0  # For visualization
pytest>=7.4.0  # For testing
matplotlib>=3.7.0 